+++ tests/Service/DanishHolidayHelperTest.php
@@ -15,7 +15,7 @@ final class DanishHolidayHelperTest extends TestCase
         $this->helper = DanishHolidayHelper::getInstance();
     }
 
-    public function testHmm()
+    public function testHmm(): void
     {
         $this->assertNotEquals(
             $this->helper->getHolidays(2024),
@@ -26,7 +26,7 @@ final class DanishHolidayHelperTest extends TestCase
     /**
      * @dataProvider dataEaster
      */
-    public function testEaster(int $year, \DateTimeInterface $expected)
+    public function testEaster(int $year, \DateTimeInterface $expected): void
     {
         $this->assertSameDate(
             $expected,
@@ -52,7 +52,7 @@ final class DanishHolidayHelperTest extends TestCase
         ];
     }
 
-    public function testHolidayNames()
+    public function testHolidayNames(): void
     {
         $year = 2024;
         $expected = [
@@ -76,7 +76,7 @@ final class DanishHolidayHelperTest extends TestCase
     /**
      * @dataProvider dataNextNonHoliday
      */
-    public function testNextNonHoliday(\DateTimeInterface $date, ?\DateTimeInterface $expected)
+    public function testNextNonHoliday(\DateTimeInterface $date, ?\DateTimeInterface $expected): void
     {
         $actual = $this->helper->getNextNonHoliday($date);
         $this->assertSameDate($expected, $actual);
@@ -93,7 +93,7 @@ final class DanishHolidayHelperTest extends TestCase
     /**
      * @dataProvider dataNextBankDay
      */
-    public function testNextBankDay(\DateTimeInterface $date, ?\DateTimeInterface $expected)
+    public function testNextBankDay(\DateTimeInterface $date, ?\DateTimeInterface $expected): void
     {
         $actual = $this->helper->getNextBankDay($date);
         $this->assertSameDate($expected, $actual);
@@ -125,7 +125,7 @@ final class DanishHolidayHelperTest extends TestCase
     /**
      * @dataProvider dataNextBankDay30
      */
-    public function testNextBankDay30(\DateTimeInterface $date, ?\DateTimeInterface $expected)
+    public function testNextBankDay30(\DateTimeInterface $date, ?\DateTimeInterface $expected): void
     {
         $actual = $this->helper->getNextBankDay($date, 30);
         $this->assertSameDate($expected, $actual);
@@ -152,7 +152,7 @@ final class DanishHolidayHelperTest extends TestCase
     /**
      * @dataProvider dataIsBankHoliday
      */
-    public function testIsBankHoliday(\DateTimeInterface $date, bool $expected)
+    public function testIsBankHoliday(\DateTimeInterface $date, bool $expected): void
     {
         $actual = $this->helper->isBankHoliday($date);
         $this->assertSame($expected, $actual);
@@ -196,10 +196,10 @@ final class DanishHolidayHelperTest extends TestCase
         ];
     }
 
-    private function assertSameDate(\DateTimeInterface $expected, \DateTimeInterface $actual, string $message = '')
+    private function assertSameDate(\DateTimeInterface $expected, \DateTimeInterface $actual, string $message = ''): void
     {
         try {
-            return $this->assertEquals($expected->getTimestamp(), $actual->getTimestamp(), $message);
+            $this->assertEquals($expected->getTimestamp(), $actual->getTimestamp(), $message);
         } catch (ExpectationFailedException $exception) {
             throw new ExpectationFailedException(sprintf('Failed asserting that %s matches expected %s.', $expected->format(\DateTimeInterface::ATOM), $actual->format(\DateTimeInterface::ATOM)), $exception->getComparisonFailure(), $exception);
         }